requirements.txt
discord.py==2.3.2
yt-dlp==2023.3.4
requests==2.31.0
spotipy==2.23.0
python-dotenv==1.0.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
psycopg2-binary==2.9.6